@@ -1,81 +1,95 @@
 [tool.poetry]
 name = "testing_homework"
-description = "Testing Homework"
+description = "Testing Homework for https://education.borshev.com/python-testing"
 version = "1.0.0"
 readme = "README.md"
-authors = ["myapp.com"]
+authors = ["Nikita Sobolev <mail@sobolevn.me>"]
+classifiers = [
+  "Private :: Do not Upload",
+]
 
 
 [tool.poetry.dependencies]
-python = "3.9.15"
+python = "3.11.5"
 
-django = { version = "^3.2", extras = ["argon2"] }
+django = { version = "^4.2", extras = ["argon2"] }
 django-split-settings = "^1.2"
-django-axes = "^5.39"
+django-axes = "^6.1"
 django-csp = "^3.7"
 django-health-check = "^3.16"
 django-http-referrer-policy = "^1.1"
-django-permissions-policy = "^4.13"
-django-stubs-ext = "^0.7"
-django-ratelimit = "^3.0"
+django-permissions-policy = "^4.17"
+django-stubs-ext = "^4.2"
+django-ratelimit = "3.0.1"  # API change in `^4.x`
 
 psycopg2-binary = "^2.9"
-gunicorn = "^20.0"
-python-decouple = "^3.6"
-structlog = "^22.1"
+gunicorn = "^21.2"
+python-decouple = "^3.8"
+structlog = "^23.1"
 requests = "^2.28"
-attrs = "^22.1"
-pydantic = "^1.10"
+attrs = "^23.1"
+pydantic = "^2.3"
 punq = "^0.6"
 
 [tool.poetry.group.dev.dependencies]
-django-debug-toolbar = "^3.6"
-django-querycount = "^0.7"
-django-migration-linter = "^4.1"
-django-extra-checks = "^0.12"
+django-debug-toolbar = "^4.2"
+django-querycount = "^0.8"
+django-migration-linter = "^5.0"
+django-extra-checks = "^0.13"
 nplusone = "^1.0"
 
-wemake-python-styleguide = "^0.17"
-flake8-pytest-style = "^1.6"
-flake8-logging-format = "^0.7"
-nitpick = "^0.32"
+wemake-python-styleguide = "^0.18"
+flake8-pytest-style = "^1.7"
+flake8-logging-format = "^0.9"
+nitpick = "^0.34"
 doc8 = "^1.0"
 
-pytest = "^7.2"
+pytest = "^7.4"
 pytest-django = "^4.5"
 pytest-cov = "^4.0"
-django-coverage-plugin = "^2.0"
-covdefaults = "^2.2.0"
-pytest-randomly = "^3.12"
+django-coverage-plugin = "^3.1"
+covdefaults = "^2.3"
+pytest-randomly = "^3.15"
 pytest-timeout = "^2.1"
-hypothesis = "^6.61"
-django-test-migrations = "^1.2"
+hypothesis = "^6.84"
+django-test-migrations = "^1.3"
 
-mypy = "^1.0"
-django-stubs = "^1.13"
+django-stubs = { version = "^4.2", extras = ["compatible-mypy"] }
 types-requests = "^2.28"
 
-yamllint = "^1.27"
-safety = "^2.1"
+djlint = "^1.32"
+yamllint = "^1.32"
+safety = "^2.3"
 dotenv-linter = "^0.4"
 polint = "^0.4"
 dennis = "^1.1"
 dump-env = "^1.3"
-ipython = "^8.4"
-import-linter = "^1.6"
+ipython = "^8.15"
+import-linter = "^1.11"
 
 [tool.poetry.group.docs]
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-sphinx = "^5.1"
-sphinx-autodoc-typehints = "^1.19"
+sphinx = "^7.2"
+sphinx-autodoc-typehints = "^1.21"
 tomli = "^2.0"
 
+
 [build-system]
-requires = ["poetry-core>=1.2.0"]
+requires = ["poetry-core>=1.6"]
 build-backend = "poetry.core.masonry.api"
 
 
+[tool.djlint]
+ignore = "H006,H030,H031,T002"
+include = "H017,H035"
+indent = 2
+blank_line_after_tag = "load,extends"
+profile = "django"
+max_line_length = 80
+format_attribute_template_tags = true
+
+
 [tool.nitpick]
-style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
+style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/0.18.0/styles/nitpick-style-wemake.toml"
