@@ -8,12 +8,10 @@ We are using `django-csp` to provide these headers.
 Docs: https://github.com/mozilla/django-csp
 """
 
-from typing import Tuple
-
 # These settings are overriden during development:
-CSP_SCRIPT_SRC: Tuple[str, ...] = ("'self'",)
-CSP_IMG_SRC: Tuple[str, ...] = ("'self'", 'https://via.placeholder.com')
-CSP_FONT_SRC: Tuple[str, ...] = ("'self'",)
-CSP_STYLE_SRC: Tuple[str, ...] = ("'self'", 'https://cdn.simplecss.org')
-CSP_DEFAULT_SRC: Tuple[str, ...] = ("'none'",)
-CSP_CONNECT_SRC: Tuple[str, ...] = ()
+CSP_SCRIPT_SRC: tuple[str, ...] = ("'self'",)
+CSP_IMG_SRC: tuple[str, ...] = ("'self'", 'https://via.placeholder.com')
+CSP_FONT_SRC: tuple[str, ...] = ("'self'",)
+CSP_STYLE_SRC: tuple[str, ...] = ("'self'", 'https://cdn.simplecss.org')
+CSP_DEFAULT_SRC: tuple[str, ...] = ("'none'",)
+CSP_CONNECT_SRC: tuple[str, ...] = ()
